Began implementing UTF-16 support in the 'Strings' operation.
