+++ src/core/config/modules/OpModules.js
@@ -18,6 +18,7 @@ import HTTPModule from "./HTTP.js";
 import ImageModule from "./Image.js";
 import JSBNModule from "./JSBN.js";
 import PublicKeyModule from "./PublicKey.js";
+import RegexModule from "./Regex.js";
 import ShellcodeModule from "./Shellcode.js";
 import URLModule from "./URL.js";
 
@@ -34,6 +35,7 @@ Object.assign(
     ImageModule,
     JSBNModule,
     PublicKeyModule,
+    RegexModule,
     ShellcodeModule,
     URLModule
 );
